Added MouseListener
Added a MouseListener to the JTable. First time using a MouseListener so I currently have it set to print out which column and row did the mouse click onto the console. Will use this to determine which CE did a User click on to pull up on the Database.

Thinking more on how Craft Essences work in the game, I might plan on changing the field types for the buffs/mlbBuffs fields within the CE class. Perhaps as an ArrayList<String>. Might be possible to even combine the fields depending on how a user enters in the data.