Added CEview() method to MainWin
Changed the types for the buffs and MLBBuffs fields of the CE.class to String. From the way CEs are set up, it's likely easier/better for a user to input Strings to describe the effects. Will likely use a multi-dimensional ArrayLists when implementing the Servant class in the future. Updated the UML to reflect the changes in the CEclass. Also updated the setters/getters to handle Strings.

On the MainWin class, added a CEview() method that is called from within the MouseListener. CEview uses a CE paramter, with the CE being retrieved from the database based on which row was clicked from within the JTable. CEview opens up a new JDialog and the CE information is added as seperate JLabels. There are 3 different sets of GridBagConstraints. Currently trying to format the JDialog so that the information is displayed uniformly. This might include changing fonts/sizes, and having the fullPortrait be somewhat scalable.