Back from hiatus
Semester is done so I'm back to being able to work on this little project.

Currently trying to create the input dialog window to create new CE. Currently building a window based off of the first early CEs, but will likely have to go back and refine this process once we start hitting the more advanced CEs that were implemented further into the game.

Added a toolbar w/ buttons to create the input dialogs to add new CEs and Servants. Once I get the CEs dialog working, I'll start with the servant dialog.