Implemented OpenFile functionality
Got the OpenFile functionality to work. This will allow the user to load in saved information (CEs for now) into the database from a save file.
Had to add a getter method to retrieve the MainWin's database field so that the mouseAdapter() had a way to refer to the database field. Will also likely need to figure out some way to display scaled images within the JTable data display. Right now, I have the table cell dimensions set to fit exactly the CE icon dimensions, but will want to change it so that the icon is scaled to fit whatever dimension the MainWin is currently at.

Next step is to figure out some way to take into account event specific effects for CEs. Will also need to start somewhere on the About dialog.

After that, start figuring out how to implement Servant Classes. Will likely need to create additional enum classes to refer to the additional advancement/upgrade materials. Ideally, will try to base code on the Atlas Academy FGO Game Data Navigator.