Quick Update
Adjusted a portion of the onAddCEClick() method to automatically fill in the MaxLVL field for the constructed CE depending on the rarity. Seeing as how it works, will probably make another change to do the same with the cost field.

Also included 3 more CEs into the database. Less than a drop compared to the entire available CE pool, but progress is progress.